build!: EXPOSED-234 Set exposed-crypt to jdk 17 & bump spring-security-crypto to 6.+
- Bump spring-security-crypto dependency from 5.8.8 to 6.2.1. As for the other
spring modules, upgrading to version 6.+ now requires JDK 17.
- Relocate tests from exposed-tests to dedicated test directory in exposed-crypt
module. exposed-tests uses JDK 8, so its dependency on the exposed-crypt project
fails.
