+++ exposed-crypt/build.gradle.kts
@@ -1,3 +1,5 @@
+import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
+
 plugins {
     kotlin("jvm") apply true
 }
@@ -7,10 +9,20 @@ repositories {
 }
 
 kotlin {
-    jvmToolchain(8)
+    jvmToolchain(17)
 }
 
 dependencies {
     api(project(":exposed-core"))
     api(libs.spring.security.crypto)
+    testImplementation(project(":exposed-dao"))
+    testImplementation(project(":exposed-tests"))
+    testImplementation(libs.junit)
+    testImplementation(kotlin("test-junit"))
+}
+
+tasks.withType<KotlinCompile>().configureEach {
+    kotlinOptions {
+        jvmTarget = "17"
+    }
 }
