+++ exposed-tests/build.gradle.kts
@@ -22,7 +22,6 @@ dependencies {
     implementation(project(":exposed-core"))
     implementation(project(":exposed-jdbc"))
     implementation(project(":exposed-dao"))
-    implementation(project(":exposed-crypt"))
 
     implementation(libs.slf4j)
     implementation(libs.log4j.slf4j.impl)
