+++ exposed-tests/src/test/kotlin/org/jetbrains/exposed/sql/tests/shared/dml/SelectTests.kt
@@ -1,9 +1,6 @@
 package org.jetbrains.exposed.sql.tests.shared.dml
 
 import nl.altindag.log.LogCaptor
-import org.jetbrains.exposed.crypt.Algorithms
-import org.jetbrains.exposed.crypt.encryptedBinary
-import org.jetbrains.exposed.crypt.encryptedVarchar
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.tests.DatabaseTestsBase
@@ -491,32 +488,6 @@ class SelectTests : DatabaseTestsBase() {
 
             val veryLongString = "1".repeat(255)
             assertEquals(0, stringTable.selectAll().where { stringTable.name eq veryLongString }.count())
-        }
-    }
-
-    @Test
-    fun testEncryptedColumnTypeWithAString() {
-        val stringTable = object : IntIdTable("StringTable") {
-            val name = encryptedVarchar("name", 80, Algorithms.AES_256_PBE_CBC("passwd", "5c0744940b5c369b"))
-            val city = encryptedBinary("city", 80, Algorithms.AES_256_PBE_GCM("passwd", "5c0744940b5c369b"))
-            val address = encryptedVarchar("address", 100, Algorithms.BLOW_FISH("key"))
-            val age = encryptedVarchar("age", 100, Algorithms.TRIPLE_DES("1".repeat(24)))
-        }
-
-        withTables(stringTable) {
-            val id1 = stringTable.insertAndGetId {
-                it[name] = "testName"
-                it[city] = "testCity".toByteArray()
-                it[address] = "testAddress"
-                it[age] = "testAge"
-            }
-
-            assertEquals(1L, stringTable.selectAll().count())
-
-            assertEquals("testName", stringTable.selectAll().where { stringTable.id eq id1 }.first()[stringTable.name])
-            assertEquals("testCity", String(stringTable.selectAll().where { stringTable.id eq id1 }.first()[stringTable.city]))
-            assertEquals("testAddress", stringTable.selectAll().where { stringTable.id eq id1 }.first()[stringTable.address])
-            assertEquals("testAge", stringTable.selectAll().where { stringTable.id eq id1 }.first()[stringTable.age])
         }
     }
 }
