+++ exposed-tests/src/test/kotlin/org/jetbrains/exposed/sql/tests/shared/dml/UpdateTests.kt
@@ -1,8 +1,5 @@
 package org.jetbrains.exposed.sql.tests.shared.dml
 
-import org.jetbrains.exposed.crypt.Algorithms
-import org.jetbrains.exposed.crypt.encryptedBinary
-import org.jetbrains.exposed.crypt.encryptedVarchar
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.dao.id.LongIdTable
 import org.jetbrains.exposed.exceptions.UnsupportedByDialectException
@@ -168,36 +165,6 @@ class UpdateTests : DatabaseTestsBase() {
                     // empty
                 }
             }
-        }
-    }
-
-    @Test
-    fun `update encryptedColumnType`() {
-        val stringTable = object : IntIdTable("StringTable") {
-            val name = encryptedVarchar("name", 100, Algorithms.AES_256_PBE_GCM("passwd", "12345678"))
-            val city = encryptedBinary("city", 100, Algorithms.AES_256_PBE_CBC("passwd", "12345678"))
-            val address = encryptedVarchar("address", 100, Algorithms.BLOW_FISH("key"))
-        }
-
-        withTables(stringTable) {
-            val id = stringTable.insertAndGetId {
-                it[name] = "TestName"
-                it[city] = "TestCity".toByteArray()
-                it[address] = "TestAddress"
-            }
-
-            val updatedName = "TestName2"
-            val updatedCity = "TestCity2"
-            val updatedAddress = "TestAddress2"
-            stringTable.update({ stringTable.id eq id }) {
-                it[name] = updatedName
-                it[city] = updatedCity.toByteArray()
-                it[address] = updatedAddress
-            }
-
-            assertEquals(updatedName, stringTable.selectAll().where { stringTable.id eq id }.single()[stringTable.name])
-            assertEquals(updatedCity, String(stringTable.selectAll().where { stringTable.id eq id }.single()[stringTable.city]))
-            assertEquals(updatedAddress, stringTable.selectAll().where { stringTable.id eq id }.single()[stringTable.address])
         }
     }
 }
